add ft_is_access validate_color_str validate_map_size fix make_rgb_array add_color clean normを意識してファイル整理
