+++ map/err_12.cub
@@ -2,7 +2,7 @@ NO ./path_to_the_north_texture
 SO ./path_to_the_south_texture
 WE ./path_to_the_west_texture
 EA ./path_to_the_east_texture
-F 220,100,0
+F 220,100,0,
 C 225,30,0
