+++ map/mini.cub
@@ -11,7 +11,7 @@ C 225,30,0
    11111111
 11110001111
 111100011
-1000000111
+1W00000111
  111111
