+++ srcs/parse_load.c
@@ -1,13 +1,13 @@
 
 #include "parse.h"
 
-void add_textures(t_loader *loader, char *str)
+static void add_textures(t_loader *loader, char *str)
 {
 	if (!is_member_default(*loader, loader->kind))
-		my_perror_exit("Duplicate elements", 0);
+		ft_perror_exit("Duplicate elements", 0);
 	else if (is_wall(loader->kind))
 		add_wall(loader, loader->kind, str);
-	else if (is_view_background(loader->kind))
+	else if (is_background(loader->kind))
 		add_color(loader, loader->kind, str);
 }
 
@@ -19,15 +19,25 @@ void	load_textures(int fd, t_loader *loader)
 	{
 		str = get_next_line(fd);
 		if (!str)
-			my_perror_exit("error1", 0);
+			ft_perror_exit("There is missing information before map_data", 0);
 		loader->kind = parse_kind(str);
 		if (loader->kind == KIND_MAP)
-			my_perror_exit("error1", 0);
+			ft_perror_exit("There is missing information before map_data", 0);
 		else if (loader->kind == KIND_FALSE)
-			my_perror_exit("Contains invalid elements", 0);
+			ft_perror_exit("Contains invalid elements", 0);
 		add_textures(loader, str);
 		free(str);
 	}
+}
+
+static void	add_map_head(t_loader *loader, char *str)
+{
+	t_map_node	*new;
+	char *val;
+
+	val = ft_strtrim_sepasets(str, "\n", " \n");
+	new = make_new_map_node(val);
+	add_last_map_node(&(loader->map_head), new);
 }
 
 void	load_map_list(int fd, t_loader *loader)
@@ -42,12 +52,10 @@ void	load_map_list(int fd, t_loader *loader)
 			break ;
 		loader->kind = parse_kind(str);
 		if (loader->kind == KIND_FALSE)
-			my_perror_exit("error3", 0);
+			ft_perror_exit("Contains invalid elements", 0);
 		if (is_texture(loader->kind))
-			my_perror_exit("error4", 0);
-		char *val = ft_strtrim_sepasets(str, "\n", " \n");
-		new = make_new_map_node(val);
-		add_last_map_node(&(loader->map_head), new);
+			ft_perror_exit("Duplicate elements", 0);
+		add_map_head(loader, str);
 		free(str);
 	}	
 }
